大修改

+++ OnlyGame/app/src/main/java/com/lifeache/gamecore/DamageCreator.java
@@ -2,7 +2,7 @@ package com.lifeache.gamecore;
 
 public class DamageCreator
 {
-	public static Damage createDamage(Unit origin,int value,int type){
+	public static Damage createDamage(Unit origin,double value,int type){
 		return new Damage(origin,value,type);
 	}
 }
